+++ cpp/arcticdb/processing/test/benchmark_clause.cpp
@@ -12,8 +12,7 @@
 #include <arcticdb/processing/clause.hpp>
 #include <arcticdb/util/test/generators.hpp>
 #include <arcticdb/column_store/memory_segment.hpp>
-#include <folly/futures/Future.h>
-#include <arcticdb/pipeline/frame_slice.hpp>
+#include <arcticdb/processing/grouper.hpp>
 
 using namespace arcticdb;
