+++ cpp/arcticdb/processing/test/test_clause.cpp
@@ -11,6 +11,7 @@
 #include <arcticdb/util/test/generators.hpp>
 #include <folly/futures/Future.h>
 #include <arcticdb/pipeline/frame_slice.hpp>
+#include <arcticdb/processing/grouper.hpp>
 
 template<typename T>
 void segment_scalar_assert_all_values_equal(const arcticdb::ProcessingUnit& proc_unit, const arcticdb::ColumnName& name, const std::unordered_set<T>& expected, size_t expected_row_count) {
